% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mat_flows.R
\name{adj_mat_flows}
\alias{adj_mat_flows}
\title{Function: adj_mat_flows()}
\usage{
adj_mat_flows(x)
}
\arguments{
\item{x}{adjacency matrix}
}
\description{
Defines internal network flows from adjacency matrix workbook
}
