% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mat_tidy.R
\name{adj_mat_tidy}
\alias{adj_mat_tidy}
\title{Tidies up adjacency matrix/matrices input}
\usage{
adj_mat_tidy(x, NLNode)
}
\arguments{
\item{x}{the sheet from network data input}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
Tidies up adjacency matrix/matrices input
}
