% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr_plot.R
\name{autocorr_plot}
\alias{autocorr_plot}
\title{autocorr_plot()}
\usage{
autocorr_plot(x, flow, lag.max = 50, addtitle = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{lag.max}{Maximum value at which to calculate acf}

\item{addtitle}{Logical argument, defaults to \code{FALSE}.
If \code{TRUE},
a title describing the type of plot is included above the plot.}

\item{...}{Other graphical parameters from \code{coda::autocorr_plot}.}
}
\value{
An autocorrelation plot over time (number of iterations).
}
\description{
a function to plot the Autocorrelation of an MCMC object
variable over time.
}
\examples{
set.seed(1)
x <- data.frame(rnorm(1000, m = 0, s = 1))
colnames(x) <- "Value"
autocorr_plot(x = x, flow = "Value")

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
autocorr_plot(x = x, flow = "Plant_GPP")

}
