% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGen_csv.R
\name{autoGen_csv}
\alias{autoGen_csv}
\title{autoGen_csv()}
\usage{
autoGen_csv(
  net_data_input,
  adj_mat_input,
  NLNode = NULL,
  respiration = TRUE,
  respiration_element = "CO2",
  primary_producer = NULL,
  author = NULL,
  date = NULL,
  weighted = TRUE,
  force = FALSE
)
}
\arguments{
\item{net_data_input}{the user-defined network input data workbook
(.xlsx) or comma separated value (.csv) filename and file extension
containing biomass and inequalities. If the argument is missing, the
function will stop and return an error prompting the user to declare
the appropriate input file.}

\item{adj_mat_input}{the user-defined input data workbook (.xlsx) or comma
separated value (.csv) filename and file extension containing
adjacency matrices of internal network flows. If the argument is missing, the
function will stop and return an error prompting the user to declare
the appropriate input file.}

\item{NLNode}{Character vector defining one or multiple non-living
compartments e.g., \code{NLNode = "Detritus"},
\code{NLNode = c("Detritus", "Detritus2")}.}

\item{respiration}{Logical. Is respiration included in the network model?
If \code{TRUE}, the default, named respiration
flows are defined from all defined living compartments (in the case of
ecological networks) to an external carbon dioxide \code{'CO2'} sink (e.g.,
\code{Invertebrate_R: Invertebrate -> CO2}). Additionally, if primary
producers are included in the network (e.g., ecological networks), Gross
Primary Production (GPP) flows are defined for each primary-producing
compartment as the flow of energy/material from the external carbon
dioxide sink to the compartment (e.g., \code{Plant_GPP: CO2 -> Plant}).
If \code{FALSE}, no
respiration flows are defined for living compartments.}

\item{respiration_element}{Character vector. If \code{NULL}, the default,
the external respiration element is defined as \code{'CO2'}.
This argument is only valid if argument \code{respiration = TRUE}.}

\item{primary_producer}{Character vector defining one or multiple
primary producer compartments (in the case of ecological networks)
e.g., \code{primary_producer = "Plant"},
\code{primary_producer = c("Plant", "Plant2")}.}

\item{author}{Character vector. LIM declaration file author name.
Defaults to system user.}

\item{date}{Character vector. Date of file creation. If \code{NULL},
the default, the date defaults to the system date.}

\item{weighted}{Logical. Return weighted LIM declaration files? If \code{NULL},
the default, both weighted and unweighted LIM declaration files are returned.
Defaults to \code{TRUE}.}

\item{force}{If set to \code{TRUE}, files of \code{data.R},
under folders \code{autoLIMR Weighted Network LIMfiles},
and \code{autoLIMR Unweighted Network LIMfiles} will be created on the
user's working directory. If this function is used in an
interactive session the user will be asked whether or not \code{data.R}
files should be created. The default value is \code{FALSE}.}
}
\value{
If argument \code{weighted = TRUE}, the default, a list of character
vectors describing a 1) weighted and 2) unweighted
network LIM declaration file. If argument \code{weighted = FALSE}, a list
containing a single LIM declaration file describing an unweighted network.
}
\description{
Reads comma separated value (CSV) files describing
1) network input data and 2) internal flows as an adjacency matrix, and
translates into a weighted and unweighted LIM declaration file compatible
with LIM and limSolve.
}
