% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGen.R
\name{autoLIMR}
\alias{autoLIMR}
\title{autoLIMR Automatically defines LIM weighted and unweighted declaration files
from two input workbooks}
\usage{
autoLIMR(
  net_data_input = "demo",
  adj_mat_input = "demo",
  NLNode = NULL,
  respiration = NULL,
  respiration_element = "CO2",
  primary_producer = NULL,
  author = NULL,
  date = NULL,
  weighted = TRUE
)
}
\arguments{
\item{net_data_input}{the user-defined workbook with network input data}

\item{adj_mat_input}{the user-defined workbook with adjacency matrices}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{respiration_element}{if Respiration = TRUE, the respiration element to
be defined. Default to "CO2"}

\item{primary_producer}{Primary producers defined in main autoLIMR function}

\item{author}{author name. Defined in the main autoLIMR function. Defaults to
system user}

\item{date}{date. Defined in the main autoLIMR function. Defaults to system
date}

\item{weighted}{whether to return weighted LIM declaration files. Default to
TRUE}
}
\value{
Two folders containing weighted, and unweighted network LIM
  declaration files respectively. For use with R package LIM
}
\description{
autoLIMR Automatically defines LIM weighted and unweighted declaration files
from two input workbooks
}
