% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_build.R
\name{check_build}
\alias{check_build}
\title{check_build(): function to check if an appropriate LIM Declaration file has
been provided}
\usage{
check_build(file)
}
\arguments{
\item{file}{R script (".R" extension) containing the LIM declaration file}
}
\description{
check_build(): function to check if an appropriate LIM Declaration file has
been provided
}
