% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultx0.R
\name{defaultx0}
\alias{defaultx0}
\title{defaultx0(): function to solve MCMC objects using the default MCMC
parsimonous solution (LIM::Xsample) as a starting point.}
\usage{
defaultx0(full_limfile, iter = NULL, jmp = NULL, x0 = NULL, ...)
}
\arguments{
\item{full_limfile}{LIM Declaration file built from check_build()}

\item{iter}{Default = NULL. Number of iterations. Default = NULL returns 3000
iterations of the MCMC samples.}

\item{jmp}{Default = NULL. If jmp = NULL, jump size is internally calculated.}

\item{x0}{Default = NULL. The inital starting point. Defaults to LSEI
(x0 = NULL), or can be changed to the Central solutions calcualted with
LIM::Xranges (x0 = "central").}

\item{...}{Further LIM::Xsample arguments.}
}
\value{
Multiple plausible network flow values. The first sample is
calculated with Least Squares with Equalities and Inequalities (LSEI).
}
\description{
defaultx0(): function to solve MCMC objects using the default MCMC
parsimonous solution (LIM::Xsample) as a starting point.
}
