% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_adj_mat.R
\name{demo_adj_mat}
\alias{demo_adj_mat}
\title{Create a named list of demo adjacency matrix sheets for 4node}
\usage{
demo_adj_mat()
}
\value{
A list containing 4 seasons worth of network input data
}
\description{
Create a named list of demo adjacency matrix sheets for 4node
}
