% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{demo_data}
\alias{demo_data}
\title{Function: demo_data()
Creates 4node demo datasets (.xlsx format)
Network input data: "autoLIMR_net_data_inputs_demo.xlsx"
Adjacency matrix input data: "autoLIMR_adj_mat_inputs_demo.xlsx"}
\usage{
demo_data(net_data_input, adj_mat_input)
}
\arguments{
\item{net_data_input}{the user-defined workbook with network input data}

\item{adj_mat_input}{the user-defined workbook with adjacency matrices}
}
\value{
Two workbooks in the working directory
}
\description{
Function: demo_data()
Creates 4node demo datasets (.xlsx format)
Network input data: "autoLIMR_net_data_inputs_demo.xlsx"
Adjacency matrix input data: "autoLIMR_adj_mat_inputs_demo.xlsx"
}
