% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fun.R
\name{export_fun}
\alias{export_fun}
\title{export_fun() part of prepack_fun() in multi_net() function}
\usage{
export_fun(x)
}
\arguments{
\item{x}{The flow matrix}
}
\value{
A matrix containing only the exports of the flow matrix
}
\description{
Grabs values of nodal exports across the boundary
}
