% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_excelLIM.R
\name{fetch_excelLIM}
\alias{fetch_excelLIM}
\title{Function fetch_excelLIM()}
\usage{
fetch_excelLIM(force = FALSE)
}
\arguments{
\item{force}{#' If set to \code{TRUE}, files of \code{excelLIM.xlsx}
will be downloaded from GitHub and saved in the users working directory. If
this function is used in an interactive session the user will be asked
whether or not \code{excelLEM.xlsx} can be downloaded. The default
value is \code{FALSE}. Users will need to explicitly change \code{TRUE} to
provide permission for download.}
}
\value{
excelLIM as .xlsx file in working directory
}
\description{
Downloads the latest version of excelLIM from GitHub
}
