% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_primary_producer_notation.R
\name{fix_primary_producer_notation}
\alias{fix_primary_producer_notation}
\title{fix_primary_producer_notation()}
\usage{
fix_primary_producer_notation(lim_files, primary_producer)
}
\arguments{
\item{lim_files}{List of LIM file contents}

\item{primary_producer}{Character vector defining primary producer compartments}
}
\value{
List of corrected LIM file contents
}
\description{
Replace Q with GPP and P with NPP for primary producers
}
