% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_write_test.R
\name{force_write_test}
\alias{force_write_test}
\title{force_write_test}
\usage{
force_write_test(data, weighted, force = FALSE, overwrite = TRUE)
}
\arguments{
\item{data}{List of character vectors describing LIM declaration files}

\item{weighted}{Logical. Whether to include weighted networks}

\item{force}{Logical. If TRUE, automatically writes files without prompting.
If NULL or not defined, shows interactive prompt.
If FALSE (default), shows interactive prompt.}

\item{overwrite}{Logical. Should existing files be overwritten? Default is TRUE.}
}
\value{
The data parameter is returned unchanged
}
\description{
For use in autoGen function. Generates interactive menu to select if files should be written to the working directory, or not.
}
\keyword{internal}
