% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_autoLIMExcel_limfile.R
\name{import_autoLIMExcel_limfile}
\alias{import_autoLIMExcel_limfile}
\title{Function: import_autoLIMExcel_limfile()#'
Imports autoLIM-Excel LIM declaration file from autoLIM-Excel}
\usage{
import_autoLIMExcel_limfile(
  file,
  weighted = TRUE,
  limname = NULL,
  open_script = FALSE
)
}
\arguments{
\item{file}{the autoLIMExcel file}

\item{weighted}{are you importing the weighted or unweighted limfile? Defaults to weighted = TRUE}

\item{limname}{the name you would like to give the network. Defaults to "Weighted_LIMfile.R" or " Unweighted_LIMfile.R" depending on weighted argument}

\item{open_script}{to view the script in R}
}
\value{
LIM Declaration File, saved as an R script
}
\description{
Function: import_autoLIMExcel_limfile()#'
Imports autoLIM-Excel LIM declaration file from autoLIM-Excel
}
\examples{
\dontrun{
import_autoLIMExcel_limfile(file = "autoLIM_Excel_test2.xlsx",
weighted = TRUE, limname = NULL, open_script = FALSE)
}
}
