% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fun.R
\name{input_fun}
\alias{input_fun}
\title{input_fun: part of prepack_net() function}
\usage{
input_fun(x)
}
\arguments{
\item{x}{an object?
the input matrix}
}
\description{
Grabs input nodes (only boundary flows in)
}
