% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM_Read.R
\name{LIM_Read}
\alias{LIM_Read}
\title{LIM_Read()}
\usage{
LIM_Read(file, verbose = FALSE, checkLinear = TRUE, remtabs = TRUE)
}
\arguments{
\item{file}{Either an object from the environment, with no quotation (e.g.,
\code{file = limfile)}, or a file from the working directory (
e.g., \code{file = 'my_limfile.R'}).}

\item{verbose}{if \code{TRUE}: prints warnings and messages to the screen.
  }

\item{checkLinear}{if \code{FALSE}: does not check for linearity
  }

\item{remtabs}{remove tabs.
  }
}
\value{
A list of objects necessary to solve the LIM problem.
}
\description{
Reads in a LIM declaration, either from an R object in the R
environment, or from a .R file. Modified from the \code{LIM::Read} function
to open the option for reading directly from the environment.
}
