% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_fun.R
\name{living_fun}
\alias{living_fun}
\title{living_fun()}
\usage{
living_fun(x)
}
\arguments{
\item{x}{Solved full flow matrix}
}
\value{
A vector of living compartments
}
\description{
Part of \code{prepack_fun()} function. Grabs living compartments.
}
