% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_def.R
\name{matrix_def}
\alias{matrix_def}
\title{Function: matrix_def()}
\usage{
matrix_def(x, mat.type)
}
\arguments{
\item{x}{input matrix}

\item{mat.type}{the matrix type to export}
}
\description{
Return matrix based on search_cols function
}
