% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diags.R
\name{mcmc_diags}
\alias{mcmc_diags}
\title{mcmc_diags(): Suite of convergence diagnostics for Markov chains.}
\usage{
mcmc_diags(x, flow = NULL, ...)
}
\arguments{
\item{x}{Object of class "multi_net_output" generated from
\code{autoLIMR::multi_net()} function.}

\item{flow}{Character argument. By default, flow = NULL, and the convergence
diagnostics are returned for each Markov Chain variable. If convergence
diagnostics are required for one Markov chain variable only,
it can be specified as a character string e.g., flow = "Plant_GPP".
We suggest leaving the default flow = NULL.}

\item{...}{Other arguments inherited from Markov chain convergence
diagnostic functions. Check defaults in
the \code{coda} package (Plummer et al., 2020).}
}
\value{
A list containing data frames of various Markov Chain
diagnostics included in package autoLIMR: 1) summary,
2) Geweke (1992) Diagnostics, 3) Effective Sample Size,
4) Heidelberger-Welch Diagnostics, 5) Raftery Lewis Diagnostics, and 6)
Autocorrelation
}
\description{
Function to return a suite of Markov Chain diagnostics
at once. Includes 1) Summary, 2) Geweke (1992) Diagnostics,
3) Effective Sample Size, 4) Heidelberger-Welch Diagnostics,
5) Raftery Lewis Diagnostics, and 6) Autocorrelation.
}
\examples{
\donttest{
fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 4000,
  jmp = 0.0001,
  x0 = NULL
)
# mcmc_diags(x = x)
mcmc_diags(x = x, flow = "Plant_GPP")
}

}
