% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{mcmc_plots}
\alias{mcmc_plots}
\title{mcmc_plots(): visual Markov chain diagnostics}
\usage{
mcmc_plots(x, flow, xranges = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{xranges}{Logical. If \code{xranges = FALSE}, the default, the lines
indicaitng minimum and maximum are excluded from the plot.
If \code{xranges = TRUE}, the plot
includes horizontal lines in the trace plot indicating absolute maximum and
minimum values of the flow range calculated by \code{LIM::Xranges}.}

\item{...}{Other base R graphical parameters. See ?plot for more details.}
}
\value{
A plot containing various Markov chain convergence diagnostic plots
for a single Markov chain variable:
1) trace, 2) kernel density, 3) running mean, 4) Geweke, and 5) Autocorrelation
}
\description{
Function to return five visual Markov chain convergence
diagnostics at once.
For any variable of MCMC object, this function returns the following plots:
1) trace,
2) kernel density,
3) running mean,
4) Geweke, and
5) Autocorrelation.
}
\examples{
fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
mcmc_plots(x = x, flow = "Plant_GPP", xranges = TRUE)
mcmc_plots(x = x, flow = "Plant_GPP", xranges = FALSE)

}
