% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sections.R
\name{merge_sections}
\alias{merge_sections}
\title{Function: merge_sections()}
\usage{
merge_sections(type = NULL, ...)
}
\arguments{
\item{type}{type of section}

\item{...}{other arguments}
}
\value{
list of LIM file sections
}
\description{
Merge lists together, and names sections for LIM declaration file
}
