% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta1.R
\name{meta1}
\alias{meta1}
\title{Function: meta1()}
\usage{
meta1(
  x,
  author = NULL,
  date = NULL,
  respiration,
  respiration_element,
  NLNode,
  weighted,
  primary_producer
)
}
\arguments{
\item{x}{network input data matrix}

\item{author}{Character vector. LIM declaration file author name.Defaults to system user.}

\item{date}{Character vector. Date of file creation. Defaults to system date.}

\item{respiration}{Logical. Is respiration included in the network model? Default = TRUE}

\item{respiration_element}{Character vector. If Respiration = TRUE,
the external respiration sink element to be defined. Defaults to "CO2"}

\item{NLNode}{Character vector defining non-living compartments e.g.,
NLNode = "Detritus", NLNode = c("Detritus", "Detritus2")}

\item{weighted}{Logical. Return weighted LIM declaration files? Defaults to TRUE}

\item{primary_producer}{Character vector defining primary producer compartments e.g.,
primary_producer = "Plant", primary_producer = c("Plant", "Plant2")}
}
\description{
Defines the first part of the metadata from network input data workbook sheets
}
