% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2.R
\name{meta2}
\alias{meta2}
\title{Function: meta2()
Defines metadata from adjacency matrices with abbreviations}
\usage{
meta2(x)
}
\arguments{
\item{x}{Adjacency matrix}
}
\description{
Function: meta2()
Defines metadata from adjacency matrices with abbreviations
}
