% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_external_list.R
\name{net_data_external_list}
\alias{net_data_external_list}
\title{Function: net_data_external_list()}
\usage{
net_data_external_list(x, respiration, respiration_element)
}
\arguments{
\item{x}{net data input matrix/matrices}

\item{respiration}{Respiration = TRUE as defined in main autoLIMR function}

\item{respiration_element}{if Respiration = TRUE, the respiration element to be defined. Default to "CO2"}
}
\description{
define externals list
}
