% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_node.R
\name{net_data_node}
\alias{net_data_node}
\title{Function: net_data_node()
define list of living/NL nodes/input/export}
\usage{
net_data_node(x, node.type, NLNode)
}
\arguments{
\item{x}{matrix/matrices}

\item{node.type}{living/nonliving/Input/Export}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
Function: net_data_node()
define list of living/NL nodes/input/export
}
