% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_resp_flows.R
\name{net_data_resp_flows}
\alias{net_data_resp_flows}
\title{Defines repiration flows}
\usage{
net_data_resp_flows(
  x,
  respiration = respiration,
  respiration_element = respiration_element,
  primary_producer = primary_producer,
  NLNode = NLNode
)
}
\arguments{
\item{x}{network data input matrix}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{respiration_element}{if Respiration = TRUE, the respiration element to be defined. Default to "CO2"}

\item{primary_producer}{Primary producers defined in main autoLIMR function}

\item{NLNode}{character non-living nodes}
}
\description{
Defines repiration flows
}
