% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLNode_mat.R
\name{NLNode_mat}
\alias{NLNode_mat}
\title{NLNode_mat
Append NLNode to non-living nodes in rows and columns of matrices}
\usage{
NLNode_mat(x, NLNode)
}
\arguments{
\item{x}{the matrix/matrices}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
NLNode_mat
Append NLNode to non-living nodes in rows and columns of matrices
}
