% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fun.R
\name{output_fun}
\alias{output_fun}
\title{part of prepack::multinets function
Grabs output}
\usage{
output_fun(x)
}
\arguments{
\item{x}{the flow matrix}
}
\description{
part of prepack::multinets function
Grabs output
}
