% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar.R
\name{pvar}
\alias{pvar}
\title{Function pvar(): Production variable definition}
\usage{
pvar(x, respiration, NLNode)
}
\arguments{
\item{x}{tidy network input data matrix}

\item{respiration}{Logical. Is respiration included in the network model? Default = TRUE}

\item{NLNode}{Character vector defining non-living compartments e.g.,
NLNode = "Detritus", NLNode = c("Detritus", "Detritus2")}
}
\description{
Defines production variables for living compartments only.
Depends on exports and arguments NLNode and respiration.
}
