% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar_w_ex.R
\name{pvar_w_ex}
\alias{pvar_w_ex}
\title{Function pvar_w_ex()
Production variable definition with exports}
\usage{
pvar_w_ex(x, respiration, NLNode)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
Function pvar_w_ex()
Production variable definition with exports
}
