% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvar.R
\name{qvar}
\alias{qvar}
\title{Function qvar(): Consumption (Q) Variable definition
Living nodes only, depends on Imports}
\usage{
qvar(x)
}
\arguments{
\item{x}{network input data matrix}
}
\description{
Function qvar(): Consumption (Q) Variable definition
Living nodes only, depends on Imports
}
