% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_sheets.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Function: read_all_sheets()
Read all excel sheets, define as list, with sheet names as list element names}
\usage{
read_all_sheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{the filename of the user-supplied workbook}

\item{tibble}{import as tibble? Always FALSE}
}
\description{
Function: read_all_sheets()
Read all excel sheets, define as list, with sheet names as list element names
}
