% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_par.R
\name{reset_par}
\alias{reset_par}
\title{reset_par(): restore default graphics parameters}
\usage{
reset_par()
}
\value{
default graphic parameters.
}
\description{
reset_par(): restore default graphics parameters
}
\details{
only default different is bg = "white" instead of bg = "transparent"
}
