% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_fun.R
\name{resp_fun}
\alias{resp_fun}
\title{resp_fun()}
\usage{
resp_fun(x)
}
\arguments{
\item{x}{The flow matrix}
}
\value{
Vector of living compartments that respire
}
\description{
Part of \code{pre_pack()} function in \code{multi_net} function.
Grabs values of internal nodes that respire.
}
