% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_notation_off.R
\name{sci_notation_off}
\alias{sci_notation_off}
\title{sci_notation_off()}
\usage{
sci_notation_off(x)
}
\arguments{
\item{x}{the imported matrix/matrices}
}
\description{
Turns off scientific notation
}
