% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssCheck.R
\name{ssCheck}
\alias{ssCheck}
\title{ssCheck}
\usage{
ssCheck(x, tol = 5, more = FALSE, zero.na = TRUE)
}
\arguments{
\item{x}{network object}

\item{tol}{tolerance}

\item{more}{not sure}

\item{zero.na}{convert NA to zero}
}
\value{
Are network objects balanced? Logical.
}
\description{
A brief description goes here
}
