% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_diag.R
\name{summary_diag}
\alias{summary_diag}
\title{summary_diag(): basic summary of autoLIMR Markov chain outputs.}
\usage{
summary_diag(x, flow = NULL, xranges = TRUE, ...)
}
\arguments{
\item{x}{Object of class "multi_net_output" generated from
autoLIMR::multi_net() function.}

\item{flow}{Character argument. By default, flow = NULL, and the convergence
diagnostics are returned for each Markov chain variable. If convergence
diagnostics are required for one Markov chain variable only,
it can be specified as a character string e.g., \code{flow = "Plant_GPP"}.
We suggest leaving the default flow = NULL.}

\item{xranges}{Logical argument, default = TRUE. Includes the LIM::Xranges
calculated flow ranges for each flow.}

\item{...}{Other arguments}
}
\value{
Data frame with each row corresponding to one Markov chain variable,
and columns describing various summary statistics.
}
\description{
Function to calculate summary statistics
for autoLIMR Markov chains.
Based on coda::summary() function (Plummer et al., 2020).
}
\examples{
# Summary of object "multi_net_output" generated from autoLIMR::multi_net().
# The example LIM Declaration files can be found in the package folder
# "example_limfiles". If using a custom LIM Declaration file, users can
# save the LIM Declaration file in the working directory and refer directly
# within the function e.g., file = "mylimfile.R".
# If the LIM Declaration file is not saved in the
# working directory, the function will accept a valid file path.

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
summary_diag(x = x) # MCMC Summary for all flows
summary_diag(x = x, flow = "Plant_GPP") # MCMC summary for flow "Plant_GPP"

}
