% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{trace_plot()}
\usage{
trace_plot(x, flow, xranges = FALSE, addtitle = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{xranges}{Logical. If \code{xranges = FALSE}, the default, the lines
indicaitng minimum and maximum are excluded from the plot.
If \code{xranges = TRUE}, the plot
includes horizontal lines in the trace plot indicating absolute maximum and
minimum values of the flow range calculated by \code{LIM::Xranges}.}

\item{addtitle}{Logical argument, defaults to \code{FALSE}.
If \code{TRUE},
a title describing the type of plot is included above the plot.}

\item{...}{Other base R graphical parameters. See ?plot for more details.}
}
\value{
A trace plot of MCMC variable over time (or number of iterations)
}
\description{
Function to plot an MCMC object variable over time (iterations)
Trace plots can be used to assess the mixing of the MCMC chain over time
}
\examples{
set.seed(1)
x <- data.frame(rnorm(1000, m = 0, s = 1))
colnames(x) <- "Value"
trace_plot(x = x, flow = "Value", xranges = FALSE)

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
trace_plot(x = x, flow = "Plant_GPP", xranges = TRUE)
trace_plot(x = x, flow = "Plant_GPP", xranges = FALSE)
}
