% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar.R
\name{uvar}
\alias{uvar}
\title{uvar()}
\usage{
uvar(x, respiration, NLNode)
}
\arguments{
\item{x}{tidy network input data matrix}

\item{respiration}{Logical. Is respiration included in the network model?
If \code{TRUE}, the default, named respiration
flows are defined from all defined living compartments (in the case of
ecological networks) to an external carbon dioxide \code{'CO2'} sink (e.g.,
\code{Invertebrate_R: Invertebrate -> CO2}). Additionally, if primary
producers are included in the network (e.g., ecological networks), Gross
Primary Production (GPP) flows are defined for each primary-producing
compartment as the flow of energy/material from the external carbon
dioxide sink to the compartment (e.g., \code{Plant_GPP: CO2 -> Plant}).
If \code{FALSE}, no
respiration flows are defined for living compartments.}

\item{NLNode}{Character vector defining one or multiple non-living
compartments e.g., \code{NLNode = "Detritus"},
\code{NLNode = c("Detritus", "Detritus2")}.}
}
\value{
Vector of U flows with vanity headings.
}
\description{
Defines unused material/energy variable for living compartments.
Depends on exports and arguments NLNode and respiration.
}
