% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar_w_ex.R
\name{uvar_w_ex}
\alias{uvar_w_ex}
\title{Function uvar_w_ex()
Unused material/energy variable definition without exports}
\usage{
uvar_w_ex(x, respiration)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}
}
\description{
Function uvar_w_ex()
Unused material/energy variable definition without exports
}
