% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar_wo_ex.R
\name{uvar_wo_ex}
\alias{uvar_wo_ex}
\title{Function uvar_wo_ex()
Unused material/energy variable definition without exports}
\usage{
uvar_wo_ex(x, respiration)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}
}
\description{
Function uvar_wo_ex()
Unused material/energy variable definition without exports
}
