% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_def.R
\name{variable_def}
\alias{variable_def}
\title{Defines whole variables sections from various other functions}
\usage{
variable_def(x, NLNode, primary_producer, respiration)
}
\arguments{
\item{x}{network input data matrix}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}

\item{primary_producer}{Primary producers defined in main autoLIMR function}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}
}
\description{
Defines whole variables sections from various other functions
}
