% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mat_ineq.R
\name{adj_mat_ineq}
\alias{adj_mat_ineq}
\title{adj_mat_ineq()}
\usage{
adj_mat_ineq(x, primary_producer = NULL, NLNode = NULL)
}
\arguments{
\item{x}{Adjacency matrix}

\item{primary_producer}{Character vector defining primary producer compartments}

\item{NLNode}{Character vector defining one or multiple non-living compartments}
}
\value{
A vector of adjacency matrix inequalities with vanity headings
to be included in the LIM declaration file
}
\description{
Defines inequalities from adjacency matrix input workbook
}
