% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aevar.R
\name{aevar}
\alias{aevar}
\title{Function aevar() : assimilation efficiency definition}
\usage{
aevar(x, respiration)
}
\arguments{
\item{x}{network data input matrix}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}
}
\description{
Defines assimilation efficiency variables if included in network input data workbook sheets
}
