% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_extended.R
\name{as_extended}
\alias{as_extended}
\title{as_extended()
Shamelessly from package enaR}
\usage{
as_extended(x, zero.na = TRUE)
}
\arguments{
\item{x}{the something to be defined}

\item{zero.na}{Change NA's to zero's}
}
\value{
Extended flow matrices
}
\description{
as_extended()
Shamelessly from package enaR
}
