% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGen.R
\name{autoGen}
\alias{autoGen}
\title{Function autoGen()}
\usage{
autoGen(
  net_data_input = "demo",
  adj_mat_input = "demo",
  NLNode = NULL,
  respiration = TRUE,
  respiration_element = "CO2",
  primary_producer = NULL,
  author = NULL,
  date = NULL,
  weighted = TRUE,
  force = FALSE
)
}
\arguments{
\item{net_data_input}{the user-defined workbook with network input data}

\item{adj_mat_input}{the user-defined workbook with adjacency matrices}

\item{NLNode}{Character vector defining non-living compartments e.g.,
NLNode = "Detritus", NLNode = c("Detritus", "Detritus2")}

\item{respiration}{Logical. Is respiration included in the network model? Default = TRUE}

\item{respiration_element}{Character vector. If Respiration = TRUE,
the external respiration sink element to be defined. Defaults to "CO2"}

\item{primary_producer}{Character vector defining primary producer compartments e.g.,
primary_producer = "Plant", primary_producer = c("Plant", "Plant2")}

\item{author}{Character vector. LIM declaration file author name.Defaults to system user.}

\item{date}{Character vector. Date of file creation. Defaults to system date.}

\item{weighted}{Logical. Return weighted LIM declaration files? Defaults to TRUE}

\item{force}{If set to \code{TRUE}, files of \code{data.R},
under folders \code{autoLIMR Weighted Network LIMfiles},
and \code{autoLIMR Unweighted Network LIMfiles} will be created on the
user's working directory. If this function is used in an
interactive session the user will be asked whether or not \code{data.R}
files should be created. The default value is \code{FALSE}.}
}
\value{
Two folders containing weighted, and unweighted network LIM
  declaration files respectively. For use with R packages 'LIM' and 'limSolve'
}
\description{
Automatically defines LIM weighted and unweighted declaration files
from two input workbooks.

This function takes some action. It also attempts to create a
few files in your working directory called \code{data.R},
under folders \code{autoLIMR Weighted Network LIMfiles},
and \code{autoLIMR Unweighted Network LIMfiles}.
If \code{data.R} files cannot be created a warning is raised.
}
\examples{
\dontrun{
autoGen(net_data_input = "your_network_data_workbook.xlsx",
adj_mat_input = "your_adjacency_matrix_data_workbook.xlsx",
NLNode = "Detritus",
primary_producer = "Plant"
respiration = TRUE,
respiration_element = "CO2",
author = "Gemma Gerber",
weighted = TRUE,
force = TRUE)}
}
