% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralx0.R
\name{centralx0}
\alias{centralx0}
\title{centralx0(): function to solve MCMC objects using the central MCMC
solution (LIM::Xsample) as a starting point.
If central solution is not valid, defaults to LSEI}
\usage{
centralx0(full_limfile, iter = NULL, jmp = NULL, x0 = "central", ...)
}
\arguments{
\item{full_limfile}{LIM Declaration file built from check_build()}

\item{iter}{Default = NULL. Number of iterations. Default = NULL returns 3000
iterations of the MCMC samples.}

\item{jmp}{Default = NULL. If jmp = NULL, jump size is internally calculated.}

\item{x0}{Default = NULL. The inital starting point. Defaults to LSEI
(x0 = NULL), or can be changed to the Central solutions calcualted with
LIM::Xranges (x0 = "central").}

\item{...}{Further LIM::Xsample arguments.}
}
\value{
Multiple plausible networks, starting point CENTRAL SOLUTION
}
\description{
centralx0(): function to solve MCMC objects using the central MCMC
solution (LIM::Xsample) as a starting point.
If central solution is not valid, defaults to LSEI
}
