% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_net_input.R
\name{demo_net_input}
\alias{demo_net_input}
\title{demo_net_input()}
\usage{
demo_net_input()
}
\value{
A list containing 4 seasons worth of network input data
}
\description{
Create a named list of demo network input data sheets for
four-node seasonal network
}
\examples{
x <- demo_net_input()

}
