% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_plot.R
\name{dens_plot}
\alias{dens_plot}
\title{dens_plot(): a function to plot the Kernel Density of an MCMC object variable}
\usage{
dens_plot(x, flow, addtitle = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{addtitle}{Logical argument, defaults to FALSE. If 'addtitles = TRUE',
a title descrbing the type of plot is included above the plot.}

\item{...}{Other graphical parameters from coda::densplot.}
}
\value{
A kernel density plot showing sample size and bin width
}
\description{
dens_plot(): a function to plot the Kernel Density of an MCMC object variable
}
\examples{
# Example 1: Kernel density plot from a data frame
# or coda::as.mcmc() object.

set.seed(1)
x <- data.frame(rnorm(1000, m = 0, s = 1))
names(x) <- "Value"
dens_plot(x = x, flow = "Value")

# Example 2: Kernel Density plot from multi_net() function output
# (class "model_class_output").
# The example LIM Declaration files can be found in the package folder
# "example_limfiles". If using a custom LIM Declaration file, users can
# save the LIM Declaration file in the working directory and refer directly
# within the function e.g., file = "mylimfile.R".
# If the LIM Declaration file is not saved in the
# working directory, the function will accept a valid file path.

fpath <- system.file("example_limfiles",
"Winter_Weighted_Network_LIMfile.R", package = "autoLIMR")
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL)
dens_plot(x = x, flow = "Plant_GPP")

}
