% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_print.R
\name{error_print}
\alias{error_print}
\title{error_print
Print Errors for undefined sheets}
\usage{
error_print(net_data_input, adj_mat_input)
}
\arguments{
\item{net_data_input}{the user-defined workbook with network input data}

\item{adj_mat_input}{the user-defined workbook with adjacency matrices}
}
\description{
error_print
Print Errors for undefined sheets
}
