% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fun.R
\name{export_fun}
\alias{export_fun}
\title{export_fun() part of prepack::multinets function
Grabs nodes that have export flows across the boundary}
\usage{
export_fun(x)
}
\arguments{
\item{x}{The flow matrix}
}
\description{
export_fun() part of prepack::multinets function
Grabs nodes that have export flows across the boundary
}
