% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_autoLIMExcel.R
\name{fetch_autoLIMExcel}
\alias{fetch_autoLIMExcel}
\title{Function: fetch_autoLIMExcel()
Downloads the latest version of autoLIM-Excel from GitHub}
\usage{
fetch_autoLIMExcel()
}
\value{
autoLIM-Excel as .xlsx file in working directory
}
\description{
Function: fetch_autoLIMExcel()
Downloads the latest version of autoLIM-Excel from GitHub
}
