% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmat_fun.R
\name{fmat_fun}
\alias{fmat_fun}
\title{fmat_fun, part of multinets function
Extracts f matrices (only internal matrices excluding boundary flows)}
\usage{
fmat_fun(x)
}
\arguments{
\item{x}{an object?}
}
\description{
fmat_fun, part of multinets function
Extracts f matrices (only internal matrices excluding boundary flows)
}
