% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmat_fun.R
\name{fmat_fun}
\alias{fmat_fun}
\title{function fmat_fun()}
\usage{
fmat_fun(x)
}
\arguments{
\item{x}{Solved full flow matrix}
}
\value{
A matrix of internal flows
}
\description{
Part of \code{pre_pack()} function in \code{multi_net()} function
Extracts flow matrices (only internal matrices excluding boundary flows)
}
\keyword{internal}
