% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke_plot.R
\name{geweke_plot}
\alias{geweke_plot}
\title{geweke_plot(): function to create Geweke (Date) plots for varibles of
MCMC objects
Used to decide...
Inspired by coda package
See (references) for more information}
\usage{
geweke_plot(
  x,
  flow,
  nbins = 20,
  pvalue = 0.05,
  frac1 = 0.1,
  frac2 = 0.5,
  addtitle = FALSE,
  ...
)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{nbins}{Number of segments.}

\item{pvalue}{p-value used to plot confidence limits for the null hypothesis.}

\item{frac1}{fraction to use from beginning of chain.}

\item{frac2}{fraction to use from end of chain.}

\item{addtitle}{Logical argument, defaults to FALSE. If 'addtitles = TRUE',
a title describing the type of plot is included above the plot.}

\item{...}{Other graphical parameters from coda::geweke.plot.}
}
\value{
Geweke plot of variable
}
\description{
geweke_plot(): function to create Geweke (Date) plots for varibles of
MCMC objects
Used to decide...
Inspired by coda package
See (references) for more information
}
\examples{
# Example 1: Geweke Plot from a data frame
# or coda::as.mcmc() object.
set.seed(1)
x <- data.frame(rnorm(1000, m = 0, s = 1))
names(x) <- "Value"
geweke_plot(x = x, flow = "Value")

# Example 2: Geweke Plot from multi_net() function output
# (class "model_class_output").
fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
geweke_plot(x = x, flow = "Plant_GPP")

}
