% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fun.R
\name{input_fun}
\alias{input_fun}
\title{input_fun() function, part of prepack_fun() function}
\usage{
input_fun(x)
}
\arguments{
\item{x}{the input matrix}
}
\description{
Grabs input nodes values (only boundary flows in)
}
