% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fun.R
\name{input_fun}
\alias{input_fun}
\title{input_fun: part of multinets function
Grabs input nodes (only boundary flows in)}
\usage{
input_fun(x)
}
\arguments{
\item{x}{an object?}
}
\description{
input_fun: part of multinets function
Grabs input nodes (only boundary flows in)
}
