% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_fun.R
\name{living_fun}
\alias{living_fun}
\title{living_fun() part of prepack::multinets function
Grabs living nodes}
\usage{
living_fun(x)
}
\arguments{
\item{x}{an object?}
}
\description{
living_fun() part of prepack::multinets function
Grabs living nodes
}
