% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_def.R
\name{matrix_def}
\alias{matrix_def}
\title{matrix_def()}
\usage{
matrix_def(x, mat_type)
}
\arguments{
\item{x}{input matrix}

\item{mat_type}{the matrix type to export}
}
\value{
a matrix of type 'Import', 'Export'
}
\description{
Return matrix based on \code{search_cols()} function
}
