% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{mcmc_plots}
\alias{mcmc_plots}
\title{mcmc_plots(): visual Markov chian doagnostics}
\usage{
mcmc_plots(x, flow, xranges = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{xranges}{Logical argument, default FALSE. If xranges = TRUE, the plot
includes horizontal lines in the trace plot indicating absolute maximum and
minimum values of the flow range calculated by LIM::Xranges.
If xranges = FALSE, the minimum and maximum lines are omitted from the plot.}

\item{...}{Other base R graphical parameters. See ?plot for more details.}
}
\value{
A plot containing various Markov chain convergence diagnostic plots
for a single Markov chian variable:
1) trace, 2) kernal density, 3) running mean, 4) Geweke, and 5) Autocorrelation
}
\description{
Function to return five visual Markov chain convergence
diagnostics at once.
For any variable of MCMC object, this function returns the following plots:
1) trace,
2) kernel density,
3) running mean,
4) Geweke, and
5) Autocorrelation.
}
\examples{
# Visual diagnostics of variables in
# Markov Chain object "multi_net_output"
# generated from autoLIMR::multi_net().

# The example LIM Declaration files can be found in the package folder
# "example_limfiles". If using a custom LIM Declaration file, users can
# save the LIM Declaration file in the working directory and refer directly
# within the function e.g., file = "mylimfile.R".
# If the LIM Declaration file is not saved in the
# working directory, the function will accept a valid file path.

fpath <- system.file("example_limfiles",
"Winter_Weighted_Network_LIMfile.R",
package = "autoLIMR")
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL)
mcmc_plots(x = x, flow = "Plant_GPP", xranges = TRUE)
mcmc_plots(x = x, flow = "Plant_GPP", xranges = FALSE)

}
