% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta1.R
\name{meta1}
\alias{meta1}
\title{Function: meta1()
Defines the first part of the metadata from network input data workbook sheets}
\usage{
meta1(
  x,
  author = NULL,
  date = NULL,
  respiration,
  respiration_element,
  NLNode,
  weighted,
  primary_producer
)
}
\arguments{
\item{x}{network input data matrix}

\item{author}{author name. Defined in the main autoLIMR function. Defaults to system user}

\item{date}{date. Defined in the main autoLIMR function. Defaults to system date}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{respiration_element}{if Respiration = TRUE, the respiration element to be defined. Default to "CO2"}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}

\item{weighted}{whether to return weighted LIM declaration files. Default to TRUE}

\item{primary_producer}{Primary producers defined in main autoLIMR function}
}
\description{
Function: meta1()
Defines the first part of the metadata from network input data workbook sheets
}
