% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_net.R
\name{multi_net}
\alias{multi_net}
\title{multi_net(): function for calculating multiple plausible networks}
\usage{
multi_net(file = NULL, iter = NULL, jmp = NULL, x0 = NULL, pack = TRUE)
}
\arguments{
\item{file}{R script (".R" extension) containing the LIM declaration file}

\item{iter}{Default = NULL. Number of iterations. Default = NULL returns 3000
iterations of the MCMC samples.}

\item{jmp}{Default = NULL. If jmp = NULL, jump size is internally calculated.}

\item{x0}{Default = NULL. The inital starting point. Defaults to LSEI
(x0 = NULL), or can be changed to the Central solutions calcualted with
LIM::Xranges (x0 = "central").}

\item{pack}{Logical. Default = FALSE. If pack = TRUE, the function returns
the flow values packed into network objects.}
}
\value{
A list of model outputs containing 1) LIM Declaration File, 2)
a table of starting solution values, 3) table of solved flow values, 4)
solved flow values packed into matrices, and if pack = TRUE, 5) list of
packed network objects, and 6) list of balanced network objects.
}
\description{
This function calculates multiple plasuible network solutions
from a LIM declaration file.
Based on LIM and limSolve packages with novel extras!
Includes options to change 1) starting points (x0), 2) jump sizes, and 3)
number of iterations. A further option, 'pack' is to pack the solved values
into network objects for network visualistion and analysis with additional
network manipulation packages (see igraph, enaR, etc.)
}
\examples{
\dontrun{
multi_net(file = file, iter = NULL, jmp = NULL, x0 = NULL, pack = FALSE)
}

}
