% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_ineq.R
\name{net_data_ineq}
\alias{net_data_ineq}
\title{Inequalities definition (with headings)}
\usage{
net_data_ineq(x, respiration, primary_producer)
}
\arguments{
\item{x}{network input data matrix}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{primary_producer}{Primary producers defined in main autoLIMR function}
}
\description{
Inequalities definition (with headings)
}
