% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_node_list.R
\name{net_data_node_list}
\alias{net_data_node_list}
\title{Function net_data_node_list()
define compartment list}
\usage{
net_data_node_list(x)
}
\arguments{
\item{x}{the net data input matrix}
}
\description{
Function net_data_node_list()
define compartment list
}
