% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_tidy.R
\name{net_data_tidy}
\alias{net_data_tidy}
\title{net_data_tidy()}
\usage{
net_data_tidy(x, NLNode)
}
\arguments{
\item{x}{Network input data sheet}

\item{NLNode}{Character vector defining one or multiple non-living
compartments e.g., \code{NLNode = "Detritus"},
\code{NLNode = c("Detritus", "Detritus2")}.}
}
\value{
Data frame of tidy input data. Tidy data includes adding 'NLNode'
strings to non-living nodes, replacing spaces " " with a period "."
in compartment names, switching off scientific notation, removing all
compartments with a biomass of  <= 0
}
\description{
Tidies up network data input sheet
}
