% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLNode_mat.R
\name{NLNode_mat}
\alias{NLNode_mat}
\title{NLNode_mat()}
\usage{
NLNode_mat(x, NLNode)
}
\arguments{
\item{x}{the matrix/matrices}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\value{
a matrix with 'NLNode' string attached to non-living compartments
}
\description{
Append NLNode to non-living nodes in rows and columns of matrices
}
