% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar.R
\name{pvar}
\alias{pvar}
\title{Function pvar(): Production variable definition
Defines production variable based on many arguments}
\usage{
pvar(x, respiration, NLNode)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
Function pvar(): Production variable definition
Defines production variable based on many arguments
}
