% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvar.R
\name{qvar}
\alias{qvar}
\title{Function qvar()}
\usage{
qvar(x)
}
\arguments{
\item{x}{tidy network input data matrix}
}
\description{
Defines consumption variables for living nodes only.
Depends on imports. For living compartments without imports, consumption
is defined as `compartment_Q = Flowto(compartment)`. For compartments with
imports, consumption is defined as `compartment_Q = Flowto(compartment) -
compartment_IM`.
}
