% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_sheets.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{read_all_sheets()}
\usage{
read_all_sheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{the file name of the user-supplied workbook}

\item{tibble}{import as tibble? Always FALSE}
}
\value{
A list of data frames, each one describing a single network defined
per sheet in the original network input data
}
\description{
Read all excel sheets, define as list,
with sheet names as list element names
}
