% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_fun.R
\name{resp_fun}
\alias{resp_fun}
\title{resp_fun() function, part of pre_pack() function in multi_net function()}
\usage{
resp_fun(x)
}
\arguments{
\item{x}{The flow matrix}
}
\description{
Grabs values of internal nodes that respire
}
