% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_fun.R
\name{resp_fun}
\alias{resp_fun}
\title{resp_fun() part of prepack::multinets function
Grabs internal, living nodes that respire}
\usage{
resp_fun(x)
}
\arguments{
\item{x}{The flow matrix}
}
\description{
resp_fun() part of prepack::multinets function
Grabs internal, living nodes that respire
}
