% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmean_plot.R
\name{runmean_plot}
\alias{runmean_plot}
\title{runmean_plot()}
\usage{
runmean_plot(x, flow, addtitle = FALSE, ...)
}
\arguments{
\item{x}{The solved MCMC flow values in a data.frame, accessible
from the multi_net() generated list of model outputs ("output") e.g.,
output[["solved.flow.values"]]. Can also refer to the entire multi_net()
output list (class(output) == "multi_net_model).}

\item{flow}{Character argument. The specified name of the flow plot.}

\item{addtitle}{Logical argument, defaults to \code{FALSE}.
If \code{TRUE},
a title describing the type of plot is included above the plot.}

\item{...}{Other base R graphical parameters. See ?plot for more details.}
}
\value{
A running mean plot over time (number of iterations).
}
\description{
Function to plot the running mean of an MCMC object
variable.
Shows time average estimates against the number of iterations.
Used to decide iteration stopping times.
}
\examples{
set.seed(1)
x <- data.frame(rnorm(1000, m = 0, s = 1))
colnames(x) <- "Value"
runmean_plot(x = x, flow = "Value")

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
runmean_plot(x = x, flow = "Plant_GPP")
}
