% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cols.R
\name{search_cols}
\alias{search_cols}
\title{search_cols()}
\usage{
search_cols(x, col_match)
}
\arguments{
\item{x}{Tidy network data input matrix}

\item{col_match}{the compartment type to match}
}
\value{
A vector of column names that match one of "Imports", "Exports",
 "Assimilation Efficiency", "Custom" or "Parameters" from the original
 network input data.
}
\description{
Search and return columns that match type (as character vector)
"Imports", "Exports", "Assimilation Efficiency"
}
