% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssCheck.R
\name{ssCheck}
\alias{ssCheck}
\title{ssCheck
A brief description goes here}
\usage{
ssCheck(x, tol = 5, more = FALSE, zero.na = TRUE)
}
\arguments{
\item{x}{not sure}

\item{tol}{not sure}

\item{more}{not sure}

\item{zero.na}{convert NA to zero}
}
\description{
ssCheck
A brief description goes here
}
