% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar.R
\name{uvar}
\alias{uvar}
\title{Function uvar()
Unused material/energy variable definition}
\usage{
uvar(x, respiration)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}
}
\description{
Function uvar()
Unused material/energy variable definition
}
