% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar.R
\name{uvar}
\alias{uvar}
\title{Function uvar()}
\usage{
uvar(x, respiration, NLNode)
}
\arguments{
\item{x}{tidy network input data matrix}

\item{respiration}{Logical. Is respiration included in the network model? Default = TRUE}

\item{NLNode}{Character vector defining non-living compartments e.g.,
NLNode = "Detritus", NLNode = c("Detritus", "Detritus2")}
}
\description{
Defines unused material/energy variable for living compartments.
Depends on exports and arguments NLNode and respiration.
}
