% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_def.R
\name{variable_def}
\alias{variable_def}
\title{variable_def()}
\usage{
variable_def(x, NLNode, primary_producer, respiration)
}
\arguments{
\item{x}{tidy network input data matrix}

\item{NLNode}{Character vector defining one or multiple non-living
compartments e.g., \code{NLNode = "Detritus"},
\code{NLNode = c("Detritus", "Detritus2")}.}

\item{primary_producer}{Character vector defining one or multiple
primary producer compartments (in the case of ecological networks)
e.g., \code{primary_producer = "Plant"},
\code{primary_producer = c("Plant", "Plant2")}.}

\item{respiration}{Logical. Is respiration included in the network model?
If \code{TRUE}, the default, named respiration
flows are defined from all defined living compartments (in the case of
ecological networks) to an external carbon dioxide \code{'CO2'} sink (e.g.,
\code{Invertebrate_R: Invertebrate -> CO2}). Additionally, if primary
producers are included in the network (e.g., ecological networks), Gross
Primary Production (GPP) flows are defined for each primary-producing
compartment as the flow of energy/material from the external carbon
dioxide sink to the compartment (e.g., \code{Plant_GPP: CO2 -> Plant}).
If \code{FALSE}, no
respiration flows are defined for living compartments.}
}
\value{
vector of variables
}
\description{
Defines whole variables sections from various other functions
}
